requirements/base.txt
asgiref>=3.10.0
Django>=5.2.9
Markdown>=3.9
packaging>=25.0
pillow>=12.0.0
psycopg2-binary>=2.9.9
pyotp>=2.9.0
python-dotenv>=1.2.1
qrcode>=8.2
sqlparse>=0.5.3
pymdown-extensions>=10.16.1
gunicorn>=21.2.0
django-ratelimit>=4.1.0
django-auditlog>=3.0.0
